src/test/javascript/spec/app/entities/materia-prima/materia-prima.service.spec.ts
import { TestBed, getTestBed } from '@angular/core/testing';
import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
import { MateriaPrimaService } from 'app/entities/materia-prima/materia-prima.service';
import { IMateriaPrima, MateriaPrima } from 'app/shared/model/materia-prima.model';

describe('Service Tests', () => {
  describe('MateriaPrima Service', () => {
    let injector: TestBed;
    let service: MateriaPrimaService;
    let httpMock: HttpTestingController;
    let elemDefault: IMateriaPrima;
    let expectedResult: IMateriaPrima | IMateriaPrima[] | boolean | null;

    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [HttpClientTestingModule],
      });
      expectedResult = null;
      injector = getTestBed();
      service = injector.get(MateriaPrimaService);
      httpMock = injector.get(HttpTestingController);

      elemDefault = new MateriaPrima(0, 'AAAAAAA', 'AAAAAAA', 'AAAAAAA');
    });

    describe('Service methods', () => {
      it('should find an element', () => {
        const returnedFromService = Object.assign({}, elemDefault);

        service.find(123).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(elemDefault);
      });

      it('should create a MateriaPrima', () => {
        const returnedFromService = Object.assign(
          {
            id: 0,
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.create(new MateriaPrima()).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'POST' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should update a MateriaPrima', () => {
        const returnedFromService = Object.assign(
          {
            tipo: 'BBBBBB',
            composicao: 'BBBBBB',
            fio: 'BBBBBB',
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.update(expected).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'PUT' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should return a list of MateriaPrima', () => {
        const returnedFromService = Object.assign(
          {
            tipo: 'BBBBBB',
            composicao: 'BBBBBB',
            fio: 'BBBBBB',
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.query().subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush([returnedFromService]);
        httpMock.verify();
        expect(expectedResult).toContainEqual(expected);
      });

      it('should delete a MateriaPrima', () => {
        service.delete(123).subscribe(resp => (expectedResult = resp.ok));

        const req = httpMock.expectOne({ method: 'DELETE' });
        req.flush({ status: 200 });
        expect(expectedResult);
      });
    });

    afterEach(() => {
      httpMock.verify();
    });
  });
});
